64bit bug fix: prevented from loading aminations
long versionId = binFile.readInt();
problem with this is when readInt returns negative int value
sign gets propagated to 64bit long value and then comparison with
CURRENT_ANIM_VERSION fails

+++ mclib/msl.cpp
@@ -2015,7 +2015,7 @@ long TG_AnimateShape::LoadTGMultiShapeAnimationFromASE (const char *fileName, TG
 			}
 			else
 			{
-				long versionId = binFile.readInt();
+				int versionId = binFile.readInt();
 				if (versionId != CURRENT_ANIM_VERSION)
 					makeBinary = true;
 			}
